requirements.txt
streamlit>=1.45.0
python-dotenv==1.0.1
pandas==2.2.2
anthropic==0.84.0
openpyxl==3.1.5